pykrs/requirements.txt
Django
django-crispy-forms
django-summernote
pytz
